+++ conformance_tests/core/test_cmdlist/src/test_cmdlist.cpp
@@ -175,7 +175,7 @@ protected:
     lzt::close_command_list(command_list);
     if (execute_all_commands) {
       lzt::execute_command_lists(command_queue, 1, &command_list, nullptr);
-      lzt::synchronize(command_queue, UINT32_MAX);
+      lzt::synchronize(command_queue, UINT64_MAX);
     }
 
     lzt::reset_command_list(command_list);
@@ -186,7 +186,7 @@ protected:
                             size, nullptr);
     lzt::close_command_list(command_list);
     lzt::execute_command_lists(command_queue, 1, &command_list, nullptr);
-    lzt::synchronize(command_queue, UINT32_MAX);
+    lzt::synchronize(command_queue, UINT64_MAX);
 
     if (execute_all_commands) {
       lzt::validate_data_pattern(host_mem, size, 1);
@@ -279,13 +279,13 @@ TEST(zeCommandListReuseTests, GivenCommandListWhenItIsExecutedItCanBeRunAgain) {
   const int num_execute = 5;
   for (int i = 0; i < num_execute; i++) {
     lzt::execute_command_lists(cmdq, 1, &cmdlist_mem_zero, nullptr);
-    lzt::synchronize(cmdq, UINT32_MAX);
+    lzt::synchronize(cmdq, UINT64_MAX);
     for (int j = 0; j < size; j++)
       ASSERT_EQ(static_cast<uint8_t *>(host_buffer)[j], 0x0)
           << "Memory Set did not match.";
 
     lzt::execute_command_lists(cmdq, 1, &cmdlist_mem_set, nullptr);
-    lzt::synchronize(cmdq, UINT32_MAX);
+    lzt::synchronize(cmdq, UINT64_MAX);
     for (int j = 0; j < size; j++)
       ASSERT_EQ(static_cast<uint8_t *>(host_buffer)[j], 0x1)
           << "Memory Set did not match.";
@@ -330,7 +330,7 @@ TEST_P(
   // Attempt to append command list after close should fail
   lzt::append_memory_set(cmdlist, buffer, &set_fail_2, size);
   lzt::execute_command_lists(cmdq, 1, &cmdlist, nullptr);
-  lzt::synchronize(cmdq, UINT32_MAX);
+  lzt::synchronize(cmdq, UINT64_MAX);
   for (size_t j = 0; j < size; j++) {
     EXPECT_EQ(static_cast<uint8_t *>(buffer)[j], set_succeed_1);
   }
@@ -345,7 +345,7 @@ TEST_P(
   // Attempt to append command list after close should fail
   lzt::append_memory_set(cmdlist, buffer, &set_fail_3, size);
   lzt::execute_command_lists(cmdq, 1, &cmdlist, nullptr);
-  lzt::synchronize(cmdq, UINT32_MAX);
+  lzt::synchronize(cmdq, UINT64_MAX);
   // No commands should be executed by command queue
   for (size_t j = 0; j < size; j++) {
     EXPECT_EQ(static_cast<uint8_t *>(buffer)[j], 0x0);
@@ -388,7 +388,7 @@ TEST_P(zeCommandListCloseAndResetTests,
     lzt::append_barrier(cmdlist, nullptr, 0, nullptr);
     lzt::close_command_list(cmdlist);
     lzt::execute_command_lists(cmdq, 1, &cmdlist, nullptr);
-    lzt::synchronize(cmdq, UINT32_MAX);
+    lzt::synchronize(cmdq, UINT64_MAX);
     for (size_t i = 0; i < test_instr; i++) {
       for (size_t j = 0; j < size; j++) {
         EXPECT_EQ(static_cast<uint8_t *>(buffer[i])[j],
@@ -409,7 +409,7 @@ TEST_P(zeCommandListCloseAndResetTests,
   }
   lzt::close_command_list(cmdlist);
   lzt::execute_command_lists(cmdq, 1, &cmdlist, nullptr);
-  lzt::synchronize(cmdq, UINT32_MAX);
+  lzt::synchronize(cmdq, UINT64_MAX);
   for (size_t i = 0; i < num_instr; i++) {
     for (size_t j = 0; j < size; j++) {
       EXPECT_EQ(static_cast<uint8_t *>(buffer[i])[j], 0x0);
@@ -454,7 +454,7 @@ TEST_P(zeCommandListFlagTests,
   lzt::append_barrier(command_list, nullptr, 0, nullptr);
   lzt::close_command_list(command_list);
   lzt::execute_command_lists(cq, 1, &command_list, nullptr);
-  lzt::synchronize(cq, UINT32_MAX);
+  lzt::synchronize(cq, UINT64_MAX);
 
   for (uint32_t i = 0; i < size; i++) {
     ASSERT_EQ(static_cast<uint8_t *>(host_memory)[i], pattern);
