+++ conformance_tests/core/test_copy/src/test_kernel_copy.cpp
@@ -79,7 +79,7 @@ TEST_P(KernelCopyTests,
 
       lzt::close_command_list(command_list);
       lzt::execute_command_lists(command_queue, 1, &command_list, nullptr);
-      lzt::synchronize(command_queue, UINT32_MAX);
+      lzt::synchronize(command_queue, UINT64_MAX);
 
       ASSERT_EQ(0, memcmp(input_data, output_data + offset,
                           (size - offset) * sizeof(int)));
@@ -174,7 +174,7 @@ TEST_P(KernelCopyTests,
 
       lzt::close_command_list(command_list);
       lzt::execute_command_lists(command_queue, 1, &command_list, nullptr);
-      lzt::synchronize(command_queue, UINT32_MAX);
+      lzt::synchronize(command_queue, UINT64_MAX);
 
       for (int i = 0; i < size; i++) {
         ASSERT_EQ(0, memcmp(input_data[i].data, output_data[i].data + offset,
